2_DEPLOY/RTUModbusModule/V2/requirements.txt
azure-iot-device==2.4.0
pymodbus
numpy
aenum
pyparsing
psycopg2-binary
sqlalchemy